[fix] 약속 만들기 시간 선택에서 초기 렌더링 문제 해결
